requirements.txt
python-telegram-bot[job-queue]==21.6
httpx==0.27.2
pydantic-settings==2.6.1
pydantic==2.9.2
aiohttp==3.10.10
